TECH: Extract effect from filters hook to prevent multiple executions
