+++ src/react/features/filters/Filters.js
@@ -1,9 +1,13 @@
 import React from 'react'
 import GraphFilter from './graph/GraphFilter'
+import MapFilter from './map/MapFilter'
+import TimelineFilter from './timeline/TimelineFilter'
 
 const Filters = () => (
   <div data-testid="filters">
     <GraphFilter />
+    <MapFilter />
+    <TimelineFilter />
   </div>
 )
