+++ src/react/features/filters/hooks/useFilter.js
@@ -1,20 +1,6 @@
-import { useEffect } from 'react'
-import { useDispatchActions, useStore } from '../../../store'
-import { activateFilter } from '../actions'
-
-const useInactivateFilterOnFocusChange = (filterId) => {
-  const [{ focus }] = useStore((state) => state.navigation)
-  const { dispatch } = useDispatchActions()
-  useEffect(() => {
-    if (focus) {
-      dispatch(activateFilter(filterId, false))
-    }
-  }, [focus, filterId])
-}
+import { useStore } from '../../../store'
 
 export const useFilter = (filterId) => {
-  useInactivateFilterOnFocusChange(filterId)
-
   const [{ filters }] = useStore((state) => state.filters)
   return filters[filterId] || {}
 }
