+++ src/react/features/filters/map/MapFilter.js
@@ -1,0 +1,8 @@
+import { useFilterEffects } from '../hooks/useFilterEffects'
+
+const MapFilter = () => {
+  useFilterEffects('map')
+  return null
+}
+
+export default MapFilter
