+++ src/react/features/filters/timeline/TimelineFilter.js
@@ -1,0 +1,8 @@
+import { useFilterEffects } from '../hooks/useFilterEffects'
+
+const TimelineFilter = () => {
+  useFilterEffects('timeline')
+  return null
+}
+
+export default TimelineFilter
